% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_trees.R
\name{agg_subtree}
\alias{agg_subtree}
\alias{scale_subtree}
\title{Aggregate or scale one subtree's children values to the parent level}
\usage{
agg_subtree(dt, by_id_cols, value_cols, col_stem, agg_function, subtree)

scale_subtree(dt, by_id_cols, value_cols, col_stem, agg_function, subtree)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Data to be aggregated or scaled.}

\item{by_id_cols}{[\code{character()}]\cr
The \code{id_cols} without the agg/scale variable(s) included.}

\item{value_cols}{[\code{character()}]\cr
Value columns that should be aggregated.}

\item{col_stem}{[\code{character(1)}]\cr
The name of the variable to be aggregated or scaled over. If aggregating an
'interval' variable should not include the '_start' or '_end' suffix.}

\item{agg_function}{[\verb{function()}]\cr
Function to use when aggregating, can be either \code{sum} (for counts) or
\code{prod} (for probabilities).}

\item{subtree}{one subtree of the list returned by \code{\link[=create_agg_subtrees]{create_agg_subtrees()}} or
\code{\link[=create_scale_subtrees]{create_scale_subtrees()}}.}
}
\value{
[\code{data.table()}] with same input columns for requested aggregates
or with scaled values for the given \code{subtree}. \code{\link[=scale_subtree]{scale_subtree()}} changes the
name of the \code{value_cols} to \verb{\{value_cols\}_scaled}.
}
\description{
Aggregate or scale one subtree's children values to the parent level
}
