% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sex.R
\name{aggregate_sex}
\alias{aggregate_sex}
\title{Aggregate sex-specific counts to both sexes combined}
\usage{
aggregate_sex(dt, id_cols, value_cols)
}
\arguments{
\item{dt}{data.table with data to be aggregated.
\itemize{
\item must only contain columns specified in \code{id_cols} and \code{value_cols}.
\item must include a column called 'sex', and only include observations for
'female' and 'male' but not 'both'.
\item each combination of \code{id_cols} (not including 'sex') must include a row for 'female' and 'male'.
\item each combination of \code{id_cols} must uniquely identify each row.
}}

\item{id_cols}{character vector of id columns that uniquely identify each row
of \code{dt}.
\itemize{
\item must include 'sex'.
}}

\item{value_cols}{character vector of value columns to be aggregated.}
}
\value{
data.table with \code{id_cols} and \code{value_cols} columns, includes additional
rows for both sexes combined.
}
\description{
Aggregate sex-specific counts to both sexes combined
}
\examples{
input_dt <- data.table::CJ(year = 2010:2020, sex = c("female", "male"),
                           value1 = 1, value2 = 2)
output_dt <- aggregate_sex(dt = input_dt, id_cols = c("year", "sex"),
                           value_cols = c("value1", "value2"))

}
