% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_scale.R
\name{assert_agg_scale_args}
\alias{assert_agg_scale_args}
\title{Check inputs to agg and scale functions}
\usage{
assert_agg_scale_args(
  dt,
  id_cols,
  value_cols,
  col_stem,
  col_type,
  mapping,
  agg_function,
  missing_dt_severity,
  overlapping_dt_severity,
  na_value_severity,
  collapse_interval_cols,
  functionality
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Data to be aggregated or scaled.}

\item{id_cols}{[\code{character()}]\cr
ID columns that uniquely identify each row of \code{dt}.}

\item{value_cols}{[\code{character()}]\cr
Value columns that should be aggregated.}

\item{col_stem}{[\code{character(1)}]\cr
The name of the variable to be aggregated or scaled over. If aggregating an
'interval' variable should not include the '_start' or '_end' suffix.}

\item{col_type}{[\code{character(1)}]\cr
The type of variable that is being aggregated or scaled over. Can be either
'categorical' or 'interval'.}

\item{mapping}{[\code{data.table()}]\cr
For 'categorical' variables, defines how different levels of the
hierarchical variable relate to each other. For aggregating 'interval'
variables, it is used to specify intervals to aggregate to, while when
scaling the mapping is inferred from the available intervals in \code{dt}.}

\item{agg_function}{[\verb{function()}]\cr
Function to use when aggregating, can be either \code{sum} (for counts) or
\code{prod} (for probabilities).}

\item{missing_dt_severity}{[\code{character(1)}]\cr
What should happen when \code{dt} is missing levels of \code{col_stem} that
prevent aggregation or scaling from occurring? Can be either 'skip',
'stop', 'warning', 'message', or 'none'. Default is 'stop'. See section on
'Severity Arguments' for more information.}

\item{overlapping_dt_severity}{[\code{character(1)}]\cr
When aggregating/scaling an interval variable or \code{collapse_interval_cols=TRUE}
what should happen when overlapping intervals are identified? Can be either
'skip', 'stop', 'warning', 'message', or 'none'. Default is 'stop'. See
section on 'Severity Arguments' for more information.}

\item{na_value_severity}{[\code{character(1)}]\cr
What should happen when 'NA' values are present in \code{value_cols}? Can be
either 'skip', 'stop', 'warning', 'message', or 'none'. Default is 'stop'.
See section on 'Severity Arguments' for more information.}

\item{collapse_interval_cols}{[\code{logical(1)}]\cr
Whether to collapse interval \code{id_cols} (not including \code{col_stem} if it is
an interval variable). Default is 'False'. If set to 'True' the interval
columns are collapsed to the most detailed common intervals and will error
out if there are overlapping intervals. See details or vignettes for more
information.}

\item{functionality}{whether \code{agg} or \code{scale} function inputs are being
checked.}
}
\value{
Invisibly returns \code{dt} but throws error if inputs are not formatted
correctly.
}
\description{
Check inputs to agg and scale functions
}
