% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_is_unique_dt.R
\name{assert_is_unique_dt}
\alias{assert_is_unique_dt}
\alias{is_unique_dt}
\title{Are the rows of the input data.table unique?}
\usage{
assert_is_unique_dt(dt, id_cols, severity = getOption("assertive.severity",
  "stop"))

is_unique_dt(dt, id_cols)
}
\arguments{
\item{dt}{input data.table to check.}

\item{id_cols}{character vector of id columns that uniquely identify each row
of \code{dt}.}

\item{severity}{How severe should the consequences of the assertion be?
Either \code{"stop"}, \code{"warning"}, \code{"message"}, or \code{"none"}.}
}
\value{
\itemize{
\item \code{is_unique_dt} returns a logical.
\item \code{assert_is_unique_dt} returns nothing but throws an error if \code{is_unique_dt}
returns FALSE.
}
}
\description{
Checks to see if the rows of the input data.table are unique for each
combination of the id columns.
}
\examples{
input_dt <- data.table::data.table(location = "France", year = 2010,
                                   sex = "female",
                                   age_start = 0:95,
                                   value1 = 2, value2 = 4)
id_cols <- c("location", "year", "sex", "age_start")
is_unique_dt(input_dt, id_cols)
assert_is_unique_dt(input_dt, id_cols)
}
