% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age_int}
\alias{calculate_age_int}
\title{Calculates the size of each age group given a start and end age for each age group}
\usage{
calculate_age_int(dt)
}
\arguments{
\item{dt}{data.table with age-specific data.
\itemize{
\item must contain numeric columns called 'age_start' and 'age_end'.
}}
}
\value{
modifies \code{dt} in place by creating a new numeric column called
'age_int'.
}
\description{
Calculates the size of each age group given a start and end age for each age group
}
\examples{
input_dt <- data.table::data.table(location = "France", year = 2010,
                                   sex = "female",
                                   age_start = 0:95,
                                   age_end = c(1:95, 125),
                                   value1 = 1, value2 = 2)
calculate_age_int(input_dt)
}
