% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age_name}
\alias{calculate_age_name}
\title{Calculate the age group name for each age group in dataset}
\usage{
calculate_age_name(dt, format = "interval", terminal_format = "plus",
  terminal_age_end = 125L)
}
\arguments{
\item{dt}{data.table with age-specific data.
\itemize{
\item must contain numeric columns called 'age_start' and 'age_end'.
}}

\item{format}{formatting style for the age group names.
\itemize{
\item must be one of 'interval', 'to', or 'dash'.
}}

\item{terminal_format}{formatting style for the terminal age group names.
\itemize{
\item only used when \code{format} is not equal to interval.
\item must be one of 'plus' or '+'.
}}

\item{terminal_age_end}{the numeric 'age_end' value for the terminal age
group.}
}
\value{
modifies \code{dt} in place by creating a new numeric column called
'age_name'.
}
\description{
Calculate the age group name for each age group in dataset
}
\examples{
input_dt <- data.table::data.table(location = "France", year = 2010,
                                   sex = "female",
                                   age_start = 0:95,
                                   age_end = c(1:95, 125),
                                   value1 = 1, value2 = 2)
calculate_age_name(input_dt, format = "interval")
}
