% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_trees.R
\name{collapse_tree}
\alias{collapse_tree}
\title{Collapse the data.tree so that intermediate nodes without data are
removed}
\usage{
collapse_tree(tree)
}
\arguments{
\item{tree}{[\code{data.tree()}]\cr
As returned by \code{\link[=create_base_tree]{create_base_tree()}} with information about where data
exists for different levels of the categorical variables.}
}
\value{
[\code{data.tree()}] with field for whether each node has data
available ('exists').
}
\description{
Collapse the data.tree so that intermediate nodes without data are
removed
}
