% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_trees.R
\name{create_interval_node}
\alias{create_interval_node}
\title{Create a node for an interval tree}
\usage{
create_interval_node(start, end, name)
}
\arguments{
\item{start}{[\code{numeric(1)}]\cr
the left endpoint of the interval tree node.}

\item{end}{[\code{numeric(1)}]\cr
the right endpoint of the interval tree node.}

\item{name}{[\code{character(1)}]\cr
name of the node in interval notation.}
}
\value{
[\code{data.tree()}] node with 'name', 'left', 'right' fields.
}
\description{
Create a node for an interval tree
}
