% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_trees.R
\name{create_agg_subtrees}
\alias{create_agg_subtrees}
\alias{create_scale_subtrees}
\title{Create a list of subtrees that can be used during aggregation or
scaling}
\usage{
create_agg_subtrees(mapping, node_exists)

create_scale_subtrees(mapping, node_exists, collapse_missing)
}
\arguments{
\item{mapping}{[\code{data.table()}]\cr
For 'categorical' variables, defines how different levels of the variable
relate to each other. For aggregating 'interval' variables, it is used to
specify intervals to aggregate to, while when scaling the mapping is
inferred from the available intervals in \code{dt}.}

\item{node_exists}{[\code{character()}]\cr
names of nodes in the mapping that data exists for.}

\item{collapse_missing}{[\code{logical(1)}]\cr
When scaling \code{categorical} data, whether to collapse missing intermediate
levels in \code{mapping}. Default is "False" and the function errors out due to
missing data.}
}
\description{
Use \code{\link[data.tree:Traverse]{data.tree::Traverse()}} to create a list of subtrees that are to be used
during aggregation or scaling.

For aggregation, traverses in post-order so that aggregation starts from
the bottom of the tree and goes up. For scaling, traverses in pre-order so
that scaling starts from the top of the tree and goes down.

Only includes nodes that can be aggregated to or scaled.
}
