% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demUtils.R
\docType{package}
\name{demUtils}
\alias{demUtils}
\alias{demUtils-package}
\title{demUtils: A package with utility functions for common demographics data and estimation problems}
\description{
The demUtils package provides two categories of important functions:
aggregation and scaling.
}
\section{aggregate functions}{

The aggregate functions aggregate counts of events by sex, age and/or location.
}

\section{scale functions}{

The scale functions scale counts by sex, age and/or location to an aggregate.
}

