% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_trees.R
\name{fill_missing_intervals}
\alias{fill_missing_intervals}
\title{create new interval nodes for any missing interval nodes}
\usage{
fill_missing_intervals(interval_tree, subtrees, col_stem)
}
\arguments{
\item{interval_tree}{[\code{data.tree()}]\cr
interval tree containing the subtree to be modified}

\item{subtrees}{[\code{list(data.tree())}]\cr
non-leaf subtrees to check and fill any missing intervals.}

\item{col_stem}{[\code{character(1)}]\cr
The name of the variable to be aggregated or scaled over. If aggregating an
'interval' variable should not include the '_start' or '_end' suffix.}
}
\value{
invisibly return reference to modified subtrees.
}
\description{
Check that the children nodes cover the entire parent interval node
and create new interval nodes for any missing ranges.
}
