% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_groupings.R
\name{identify_unique_groupings}
\alias{identify_unique_groupings}
\title{Determine unique sets of the agg/scale variable present in the input
data.table}
\usage{
identify_unique_groupings(
  dt,
  id_cols,
  col_stem,
  col_type,
  collapse_interval_cols
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Data to be aggregated or scaled.}

\item{id_cols}{[\code{character()}]\cr
ID columns that uniquely identify each row of \code{dt}.}

\item{col_stem}{[\code{character(1)}]\cr
The name of the variable to be aggregated or scaled over. If aggregating an
'interval' variable should not include the '_start' or '_end' suffix.}

\item{col_type}{[\code{character(1)}]\cr
The type of variable that is being aggregated or scaled over. Can be either
'categorical' or 'interval'.}

\item{collapse_interval_cols}{[\code{logical(1)}]\cr
Whether to collapse interval \code{id_cols} (not including \code{col_stem} if it is
an interval variable). Default is 'False'. If set to 'True' the interval
columns are collapsed to the most detailed common intervals and will error
out if there are overlapping intervals. See details or vignettes for more
information.}
}
\value{
list with two named elements.
\enumerate{
\item \emph{groupings}: [\code{data.table()}] with \code{id_cols} (except \code{col_stem}
columns) and a new column 'available_vars' that is a ';' separated string
for each unique variable present in each combination of \code{id_cols} (except
\code{col_stem} columns). If \code{col_type} is 'interval' then each 'start' and
'end' variable is also ',' separated.
\item \emph{unique_groupings}: [\code{data.table()}] with only one common for the
unique 'available_vars' in the \code{groupings}.
}
}
\description{
Determine unique sets of the agg/scale variable present in the input
data.table
}
\seealso{
\code{\link[=subset_unique_grouping]{subset_unique_grouping()}}
}
