% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iran_mapping}
\alias{iran_mapping}
\title{Iran Provinces Mapping from 1950 to Today}
\format{A data.table with 47 rows and 2 columns:
\itemize{
\item parent: the aggregate location.
\item child: one of the constituent locations.
}}
\usage{
iran_mapping
}
\description{
An example mapping of Iran and its provinces from 1950 to
present day. The mapping accounts for province boundary changes over time.
}
\details{
Each row in the data.table represents the relationship between Iran, its
historical provinces, and present day provinces.

This mapping can be used for aggregating \code{\link[=agg]{agg()}} or scaling \code{\link[=scale]{scale()}} values.
}
\seealso{
\code{\link[=create_agg_tree]{create_agg_tree()}}, \code{\link[=create_scale_tree]{create_scale_tree()}}, \code{\link[=vis_tree]{vis_tree()}}
}
\keyword{datasets}
