% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iran_pop}
\alias{iran_pop}
\title{Iran Census Population Counts}
\format{A data.table with 10764 rows and 9 columns:
\itemize{
\item location: name of the location (corresponds to a location in
\code{location_mapping}).
\item year: year the census was conducted.
\item sex: either 'female' or 'male'.
\item age_start: start of the age interval.
\item age_end: end of the age interval.
\item source_name: name of the source that provided the census data population
count.
\item nid: unique identifier for the catalog record in the Global Health Data
Exchange (GHDx) that directly provided the population count.
\item underlying_nid: unique identifier for the catalog record in the Global
Health Data Exchange (GHDx) that the source can be attributed to if
provided by a data vendor (like the Demographic Year Book, DYB).
\item population: population count for the location-sex-age_group-source.
}}
\usage{
iran_pop
}
\description{
Publicly available sex and age specific population counts for
Iran at the national level and provinces that existed during the census year.
}
\keyword{datasets}
