% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_utils.R
\name{logit}
\alias{logit}
\alias{invlogit}
\title{Logit and inverse logit functions}
\usage{
logit(x, scale = 1)

invlogit(x, scale = 1)
}
\arguments{
\item{x}{[\code{numeric()}]\cr Value to take the logit or inverse logit of.}

\item{scale}{[\code{numeric()}]\cr Value to scale by.}
}
\value{
[\code{numeric()}]\cr The calculated logit or inverse logit
value.
}
\description{
Logit and inverse logit functions
}
\details{
\itemize{
\item logit (x) = log (x / (1-x))
\item inverse logit (x) = exp (x) / (1 + exp (x))
}

Optional scaled logit with 'scale' argument:
\itemize{
\item logit (x, s) = s * log (x / (1-x))
\item inverse logit (x, s) = exp (x/s) / (1 + exp (x/s))
}
}
\examples{
logit(0.1)
invlogit(-2)

}
