% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_formatting.R
\name{name_to_start_end}
\alias{name_to_start_end}
\title{Parse interval notation name to interval endpoints}
\usage{
name_to_start_end(name)
}
\arguments{
\item{name}{[\code{character()}]\cr
left-closed, right-open interval notation as described in `gen_name().}
}
\value{
[\code{list()}] with element for 'start' containing a numeric vector of
left-closed endpoints and another element for 'end' containing a numeric
vector of right-opens endpoints.
}
\description{
Parse interval notation name to interval endpoints
}
