% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{place_new_interval_node}
\alias{place_new_interval_node}
\title{Place a new interval node in an interval tree}
\usage{
place_new_interval_node(current_node, new_node)
}
\arguments{
\item{current_node}{[\code{data.tree()}]\cr
node of interval tree to put the \code{new_node} in.}

\item{new_node}{[\code{data.tree()}]\cr
new interval node to place somewhere below \code{current_node}.}
}
\value{
Invisibly returns reference to modified \code{current_node} with
\code{new_node} placed as part of subtree.
}
\description{
Recursive function to place a new interval node in an interval
tree.
}
\details{
Assumption is that \code{new_node} is a sub interval of \code{current_node} and
this is double checked.

\code{new_node} can be placed below \code{current_node} as:
\itemize{
\item another child node of \code{current_node}.
\item another child node of \code{current_node} but with one of \code{current_node}'s
children placed as a child of \code{new_node}.
\item somewhere below one of \code{current_node}'s children.
}
}
