% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_trees.R
\name{identify_missing_agg}
\alias{identify_missing_agg}
\alias{identify_missing_scale}
\alias{identify_present_agg}
\title{Identify names of problematic nodes in categorical trees}
\usage{
identify_missing_agg(tree)

identify_missing_scale(tree)

identify_present_agg(tree)
}
\arguments{
\item{tree}{[\code{data.tree()}] as returned by \code{\link[=create_agg_tree]{create_agg_tree()}} or
\code{\link[=create_scale_tree]{create_scale_tree()}}.}
}
\value{
[\code{character()}] vector with the names of the problematic nodes.
}
\description{
Identify names of problematic nodes that are making aggregation
or scaling not possible because they are missing or the aggregates are
already present.
}
