% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_groupings.R
\name{subset_unique_grouping}
\alias{subset_unique_grouping}
\title{Subset to data for one unique set of the agg/scale variable}
\usage{
subset_unique_grouping(
  dt,
  id_cols,
  col_stem,
  col_type,
  collapse_interval_cols,
  groups,
  group_num
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Data to be aggregated or scaled.}

\item{id_cols}{[\code{character()}]\cr
ID columns that uniquely identify each row of \code{dt}.}

\item{col_stem}{[\code{character(1)}]\cr
The name of the variable to be aggregated or scaled over. If aggregating an
'interval' variable should not include the '_start' or '_end' suffix.}

\item{col_type}{[\code{character(1)}]\cr
The type of variable that is being aggregated or scaled over. Can be either
'categorical' or 'interval'.}

\item{collapse_interval_cols}{[\code{logical(1)}]\cr
Whether to collapse interval \code{id_cols} (not including \code{col_stem} if it is
an interval variable). Default is 'False'. If set to 'True' the interval
columns are collapsed to the most detailed common intervals and will error
out if there are overlapping intervals. See details or vignettes for more
information.}

\item{groups}{[\code{list(2)}]\cr
Output of \code{\link[=identify_unique_groupings]{identify_unique_groupings()}}.}

\item{group_num}{[\code{integer(1)}]\cr
Integer row of \code{groups$unique_groupings} to subset data to.}
}
\value{
list with two named elements.
\enumerate{
\item \emph{unique_cols}: [\code{data.table()}] with unique combinations of the
variable being aggregated/scaled. If \code{col_type} is 'categorical' then just
a column for \code{col_stem}. If \code{col_type} is interval' then a column for
'\code{{col_stem}}_start' and '\code{{col_stem}}_end'
then the 'start' and 'end' variable values are also ',' separated.
\item \emph{dt}: [\code{data.table()}] subset of the input \code{dt} that has a specific
combination of the variable being aggregated or scaled over.
}
}
\description{
Subset to data for one unique set of the agg/scale variable
}
\seealso{
\code{\link[=identify_unique_groupings]{identify_unique_groupings()}}
}
