% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_draws.R
\name{summarize_dt}
\alias{summarize_dt}
\title{Calculate summary statistics}
\usage{
summarize_dt(
  dt,
  id_cols,
  summarize_cols,
  value_col,
  summary_fun = c("mean"),
  probs = c(0.025, 0.975)
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Data to calculate summary statistics for.}

\item{id_cols}{[\code{character()}]\cr
ID columns that uniquely identify each row of \code{dt}.}

\item{summarize_cols}{[\code{character()}]\cr
The \code{id_cols} that should be collapsed and to calculate summary statistics
over.}

\item{value_col}{[\code{character(1)}]\cr
Value column that summary statistics should be calculated for.}

\item{summary_fun}{[\code{character()}]\cr
Names of the functions that can be used to summarize a vector of values.
Default is "mean".}

\item{probs}{[\code{numeric()}]\cr
Probabilities with values in \verb{[0,1]} to be used when producing sample
quantiles with \code{stats::quantile()}. Default is the 2.5 and 97.5 quantiles.}
}
\value{
[\code{data.table()}] with \code{id_cols} (minus the \code{summarize_cols}) plus
summary statistic columns. The summary statistic columns have the same name
as each function specified in \code{summary_fun} and the quantiles are named
like 'q_\code{(probs * 100)}'.
}
\description{
Calculate summary statistics when collapsing over a certain
variable. For example can calculate summary statistics across a set of
draws, or locations, or location-years, etc.
}
\details{
\code{summary_fun} correspond to names of functions in R that can take a vector of
values and reduce to one summary statistic. This can also include user
defined functions specified in the global environment.

The \code{probs} argument is used to specify the probabilities to calculate sample
quantiles for using the \code{stats::quantile()} function. The default 2.5 and
97.5 quantiles would have columns named 'q2.5' and 'q97.5' in the returned
data.table.
}
\examples{
input_dt <- data.table::data.table(location = "USA",
                                   draw = 1:101,
                                   value = 0:100)
output_dt <- summarize_dt(dt = input_dt,
                          id_cols = c("location", "draw"),
                          summarize_cols = "draw",
                          value_col = "value")
}
