% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_trees.R
\name{agg_subtree}
\alias{agg_subtree}
\alias{scale_subtree}
\title{Aggregate or scale one subtree's children values to the parent level}
\usage{
agg_subtree(
  dt,
  id_cols,
  value_cols,
  col_stem,
  col_type,
  agg_function,
  subtree,
  missing_dt_severity,
  overlapping_dt_severity,
  collapse_interval_cols
)

scale_subtree(
  dt,
  id_cols,
  value_cols,
  col_stem,
  col_type,
  agg_function,
  subtree,
  missing_dt_severity,
  overlapping_dt_severity,
  collapse_interval_cols
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Data to be aggregated or scaled.}

\item{id_cols}{[\code{character()}]\cr
ID columns that uniquely identify each row of \code{dt}.}

\item{value_cols}{[\code{character()}]\cr
Value columns that should be aggregated.}

\item{col_stem}{[\code{character(1)}]\cr
The name of the variable to be aggregated or scaled over. If aggregating an
'interval' variable should not include the '_start' or '_end' suffix.}

\item{col_type}{[\code{character(1)}]\cr
The type of variable that is being aggregated or scaled over. Can be either
'categorical' or 'interval'.}

\item{agg_function}{[\verb{function()}]\cr
Function to use when aggregating, can be either \code{sum} (for counts) or
\code{prod} (for probabilities).}

\item{subtree}{one subtree of the list returned by \code{\link[=create_agg_subtrees]{create_agg_subtrees()}} or
\code{\link[=create_scale_subtrees]{create_scale_subtrees()}}.}

\item{missing_dt_severity}{[\code{character(1)}]\cr
What should happen when \code{dt} is missing levels of \code{col_stem} that
prevent aggregation or scaling from occurring? Can be either 'skip',
'stop', 'warning', 'message', or 'none'. Default is 'stop'. See section on
'Severity Arguments' for more information.}

\item{overlapping_dt_severity}{[\code{character(1)}]\cr
When aggregating/scaling an interval variable or \code{collapse_interval_cols=TRUE}
what should happen when overlapping intervals are identified? Can be either
'skip', 'stop', 'warning', 'message', or 'none'. Default is 'stop'. See
section on 'Severity Arguments' for more information.}

\item{collapse_interval_cols}{[\code{logical(1)}]\cr
Whether to collapse interval \code{id_cols} (not including \code{col_stem} if it is
an interval variable). Default is 'False'. If set to 'True' the interval
columns are collapsed to the most detailed common intervals and will error
out if there are overlapping intervals. See details or vignettes for more
information.}
}
\value{
[\code{data.table()}] with same input columns for requested aggregates
or with scaled values for the given \code{subtree}. \code{\link[=scale_subtree]{scale_subtree()}} changes the
name of the \code{value_cols} to \verb{\{value_cols\}_scaled}.
}
\description{
Aggregate or scale one subtree's children values to the parent level
}
